Add legend and create show and export methods
You can now add frontside legends to your caps! The reason for frontside
is to be able to print it with two colours when your printer only
supports one. You first print the legend, then switch colours, and
without removing the legend you print the cap over it.

Added the methods show and export to make it easier to get the right
orientation of the caps and maybe be more friendly to users. You can
still just do the build and do your own thing ofc. But it seems like a
nice to have.