Fix support for RIGHT and CENTER stepped keys
You can now specify a whole lot more regarding the stepped keys. How
wide the raised part should be, how high the step should be, if the
raised part shold be on the left, center, or right. Pretty fun.

Have not done work with stepped ISO enter. That is a nother can of
worms to do.