Fix legends, reduce fillet, up export resolution
The legends did not play nice. The problem was that each cap looked very
different from one another after filleting. So, place the legend based
on the non-filleted cap. But we can't add the legend before filleting as
that messes up other things. Thankfully we can measure the placement of
the legend on the non processed cap and transfer that position to the
processed one.
Also added so you can manually select which face you want the legend on.

0.6 caused problems with long caps. No idea why. Reduce to 0.5 as the
default. Not much changes but we still fillet to make the caps nicer.

The default resolution for export was bad. The curves were not nice on
iso enter. So reduce the tolerance to something better, upping the
resolution of the cap.