Refactor from single file to module (#1)
* Refactor into to more appropriate structure

Put things in different places so it's easier to expand and contain
things to what they need. One massive file is just a pain to work with.

* Fix dish cutter for inverted caps

* Fix legends, reduce fillet, up export resolution

The legends did not play nice. The problem was that each cap looked very
different from one another after filleting. So, place the legend based
on the non-filleted cap. But we can't add the legend before filleting as
that messes up other things. Thankfully we can measure the placement of
the legend on the non processed cap and transfer that position to the
processed one.
Also added so you can manually select which face you want the legend on.

0.6 caused problems with long caps. No idea why. Reduce to 0.5 as the
default. Not much changes but we still fillet to make the caps nicer.

The default resolution for export was bad. The curves were not nice on
iso enter. So reduce the tolerance to something better, upping the
resolution of the cap.

* Fix thickness

* Move Legend to subpackage

Break out adding of legend to own files

* Move dish to own file

* Break out homing

* Rewrite stems and supports

You can now go real wild and crazy with both stems and their supports.
Instead of relying on some calculations that will eventually be wrong,
extrude things using until="next". This will create a support that will
always reach the next possible edge, making long keys able to print on
the same edge as wide keys. Much fun! It means ISO Enter can have its
legend facing the user instead of the side.

* No need to rotate stems for ISO enter anymore

* Fix Z delta for supports

* Re-arrange and remove odd selection

* General fix

* Fix legends

* Fix clone